Use a context capable of injectin frame priority if stream class is A or B
This allows frames to be properly marked with PCP so that both the
network stack on the card as well as the network can pick it up and
handle it appropriately.

NOTE: if enabled, an AVB-capable switch will recognize this as a proper
AVB frame and will drop it if no reservation exists.