Do not clear data after send.
We can send duplicates of the data because data from each sensor is
time-stamped. This allows the receiver to detect duplicate data.

The alternative of clearing the data set will result in a 0 value being
sent, which could be argued is the same thing (the value being invalid),
by clearing it, we are effectively forcing the receiver to act, whereas
we now let the receiver decide. If they are happy with old data, then
code complexity will be reduced.